tests/wpt/tests/css/css-text/parsing/text-autospace-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-autospace with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#propdef-text-autospace">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-autospace", "normal");
test_valid_value("text-autospace", "no-autospace");

test_valid_value("text-autospace", "auto");

test_valid_value("text-autospace", "ideograph-alpha");
test_valid_value("text-autospace", "ideograph-numeric");
test_valid_value("text-autospace", "ideograph-alpha ideograph-numeric");

test_valid_value("text-autospace", "punctuation");
test_valid_value("text-autospace", "punctuation normal");
test_valid_value("text-autospace", "punctuation ideograph-alpha");
test_valid_value("text-autospace", "punctuation ideograph-alpha ideograph-numeric");

test_valid_value("text-autospace", "ideograph-alpha insert");
test_valid_value("text-autospace", "ideograph-numeric insert");
test_valid_value("text-autospace", "ideograph-alpha ideograph-numeric insert");

test_valid_value("text-autospace", "punctuation insert");
test_valid_value("text-autospace", "punctuation normal insert");
test_valid_value("text-autospace", "punctuation ideograph-alpha insert");

test_valid_value("text-autospace", "insert ideograph-alpha");
test_valid_value("text-autospace", "insert ideograph-numeric");
test_valid_value("text-autospace", "insert ideograph-alpha ideograph-numeric");

test_valid_value("text-autospace", "insert punctuation");
test_valid_value("text-autospace", "insert punctuation normal");
test_valid_value("text-autospace", "insert punctuation ideograph-alpha");
test_valid_value("text-autospace", "insert punctuation ideograph-alpha ideograph-numeric");

test_valid_value("text-autospace", "ideograph-alpha replace");
test_valid_value("text-autospace", "ideograph-numeric replace");
test_valid_value("text-autospace", "ideograph-alpha ideograph-numeric replace");

test_valid_value("text-autospace", "punctuation replace");
test_valid_value("text-autospace", "punctuation normal replace");
test_valid_value("text-autospace", "punctuation ideograph-alpha replace");

test_valid_value("text-autospace", "replace ideograph-alpha");
test_valid_value("text-autospace", "replace ideograph-numeric");
test_valid_value("text-autospace", "replace ideograph-alpha ideograph-numeric");

test_valid_value("text-autospace", "replace punctuation");
test_valid_value("text-autospace", "replace punctuation normal");
test_valid_value("text-autospace", "replace punctuation ideograph-alpha");
test_valid_value("text-autospace", "replace punctuation ideograph-alpha ideograph-numeric");
</script>
</body>
</html>
